add healthcheck to lexbox-api and fix docker compose hasura dep
Previously on a clean init of hasura and lexbox-api there was a race condition where hasura depends on the existence of the db schema created on init by the lexbox-api service.  This change adds a healthcheck and re-orders the docker compose dependencies so that lexbox-api starts first, and then hasura starts up to do it's initialization, having confidence the schema exists when it starts up.