Keep track of revisions to skip in `git blame`
Any revision that touches a large number of files, like the recent
commit that removes BOM marks from the start of many, many files, will
cause `git blame` to find that that file was touched. We can tell git
blame to skip certain revs by passing a filename, like so:

    git blame --ignore-revs-file .git-blame-ignore-revs

Or we could run the following command to use that file by default:

    git config --global blame.ignoreRevsFile .git-blame-ignore-revs

Once this PR is merged, the ignore-revs file will also need the hash of
the merge commit added to it, becuase that commit will also cause git
blame to think the files were touched. It's impossible to do that ahead
of time because you cna't know the hash until the commit is created.