Settle on using fromColor in train tracks
After considering whether the stroke color for a train-tracks path
should match the circle above it or below it (only matters on a fork or
merge commit), I settled on matching the color above it. That means that
a fort will show two different colors branching off from it, which is
really what you want visually so that you can see two different lines of
development starting from that point.

Also removed the comment suggesting considering toColor, as we decided
not to do that.