Add SIL.Chorus.Mercurial package to get hg binary
The package copies the Mercurial and MercurialExtensions folders into a
directory of your choice. We will put it in the project root, and copy
it into the output directory, so that it can refer to `Mercurial/hg` in
the source and it will work both in development and when deployed.