Run LexBoxApi dev container as UID 1000
This avoids issues with file permissions. Without this the build output
is being created as root, so root-owned files end up in the developer's
local filesystem, making it impossible to run `dotnet build` outside the
container. This is only an issue for developers running Linux, as devs
running Windows won't encounter the issue due to Windows's different
permissions model.