Add FixFwData.exe, combine clone and S/R tests
FixFwData.exe doesn't need to actually do anything for the purpose of
these basic tests. The Send/Receive test needs to run *after* the clone
test, and the simplest way to do that at the moment is to put them both
in the same unit test (since xUnit executes tests in parallel).

Tests are currently failing due to a permissions error in hgweb: the
hgweb process is run under user www-data, but the volume is created with
ownership by the user ID who owns the /src folder (usually UID 1000). We
cna get around that by deleting and re-creating the hgweb volume after
modifying the Dockerfile for hgweb.