Use modern Mercurial (6.x) in hgresumable image
This should ensure that the repos we clone (which use the share-safe
format) are readable by hgresumable as well as hgweb.

Note that we delete the system-installed Mercurial too, so that the one
in /usr/local/bin/hg will be the only one running and /usr/bin/hg won't
accidentally run instead.