More Send/Receive testing (#101)
We now send valid *and* invalid credentials both conventional and
resumable HG servers, allowing us to test all the combinations we need
tested.

The following env vars will be used to define the hostnames:
- TEST_STANDARD_HG_HOSTNAME (default "hg.localhost")
- TEST_RESUMABLE_HG_HOSTNAME (default "resumable.localhost")

When running locally, `dotnet test -e NAME=VALUE` can be used to define
environment variables for that test run. In GitHub Actions, the `env:`
block is the best approach.