Better login via load functions (#123)
The preferred approach for keeping track of users in Svelte-Kit is as follows:

- The `handle` hook adds a `getUser` function to `event.locals`, which
  will be called later on when user info is needed
    - The function closes over `event.cookies`, so it can be called
      later on without parameters
- The root layout, in the **server** `load` function, calls `getUser`
  and puts user info into its return value
- Any load function that needs user data can do `const { user } = await
  event.parent()`
- Any *page* that needs user data can access `$page.data.user`
- Most important: if someone signs out, the load function re-runs itself
  so that `$page.data.user` is always up-to-date and doesn't end up
  getting stale

This PR modifies the login logic to follow that algorithm.

* Fix login process - use Svelte-Kit load properly

Now the login process will redirect to a server page, forcing a cookie
reload so that the rest of the app will correctly get the user from the
load() function in src/routes. This will mean that the user store should
no longer be stale.

* Fix SHA1 hashing during SSR

The `crypto.subtle` module is a browser-only module; if we're doing
server-side rendering then we need the NodeJS equivalent.

* Get rid of user store, use page data instead

Now the user store can be accessed from `data.user` in pages, or
`$page.data.user` in any components.

* Allow refreshing user data in pages if needed

Any page needing to refresh user data can call invalidate('user:(id)')
and that will trigger this load function to re-run, keeping the page
data up-to-date. This will be useful later, such as once the "edit user
data" PR gets merged in.