Greatly increase memory request for dotnet watch
The dotnet watch command is memory-hungry, and if the memory request is
too low, the container gets OOM killed and then the new dotnet watch has
to re-restore all the packages, which takes a while. Upping the memory
limit means that dotnet watch is able to do its job properly.