Hasura no longer needs to wait on backend
Now that the database is persistent, the only time Hasura will give an
error is the first time you spin up the whole system, when the database
is completely empty. And even there, we may be able to avoid that by
having an initContainer to check that `SELECT * FROM Projects LIMIT 1`
returns one line rather than zero, which means that the database has
been populated by the backend's SeedingData.SeedDatabase task.