Have Vite listen on all interfaces
Vite only listens on the loopback interface by default, which is fine
for local development but doesn't work in a Kubernetes container, where
the container is assigned its own cluster-local IP address that you're
supposed to listen to. Telling vite to listen to 0.0.0.0 means "all
available interfaces" which is what we want for k8s dev.