Handle case where project had no repo dir
If the repo directory cannot be found, HgService.BackupRepo will return
null for the backup file path, which will turn into a 404 in the
ProjectController. We'll want to detect that 404 in the UI when the user
clicks on the "Backup Project" button and disallow submitting the form
if that happens.