Attempt to fix "subscribe to page store" error
For some reason, the layout code is sometimes throwing an error about
subscribing to the Svelte `$page` store outside of component
initialization. It's unclear how that's happening as the only place
we're using `$page` is in the component init, but let's try using
Svelte's `getStores()` function to avoid that delayed subscription.