Add API to update multiple unknown projects at once
Query parameters limit and offset can be used to limit how much work
we're asking hgweb to do at once; default limit is 50. The offset is
there so that if there are multiple projects whose type we can't
determine, and we know how many there are, we can choose to skip them
next time we run the operation.