Use chg for greater speed at the cost of some RAM
The chg command is implemented in C for faster startup speed (no need to
start a Python interpreter each time). It starts up a Python hg process
and keeps it running as a daemon, then talks to it over a Unix socket.
So the second and subsequent times you run chg, you don't have to pay
the Python startup cost. Worth it if we'll run lexentrycount often.