Add `hg tip` command for easier caching
This might be easily obtainable from hgweb, but it's probably faster to
use the command runner to do it. We'll want to cache the SHA of the
repo's most recent commit, and only update the LexEntry count if a
Send/Receive process leaves the repo with a new SHA at the tip.