Change entry count to work in SSR
Now the only situation where the entry count won't show up in
server-side rendering is if it's not in the GQL data and we have to make
a second API call for it. In that case, the API call will return a
promise and the client-side code will await it. In all other cases, the
server-side code will be able to render with the GQL results.