Better UI for project manager inviting users by email
Adding a user to a project, if the email address did not yet exist in
Lexbox, now succeeds and closes the dialog box. The notification that
shows up is different, however: instead of saying "(email) has been
added to the project", it says "(email) has been sent an invitation
email to join the project". Then the project membership doesn't show up
until the emailed link has been clicked on and the user registration
process completed for that new user.