Put test ID in temp project name
If two copies of the test run simultaneously in Chromium and Firefox,
you can end up with the situation where one of them sees two copies of
the test project in the admin dashboard. And while they have different
project codes, until now they had the same project name. However, since
the admin dashboard code to click on the project uses the project name
(with `{exact: true}`), the locator was seeing two copies of the
project, and therefore throwing an error.

This commit fixes that by putting the test ID (which is different for
each browser) into the project name as well as the project code, so that
the admin dashboard locators will only ever find a single project.