Stop Playwright complaining about fixture
Playwright complains if you don't use the object destructuring pattern
on the first argument of a fixture, even if it doesn't need any
dependencies passed to it. So we pass an empty object as the first
parameter of the tempDir fixture, which then requires telling eslint to
stop complaining about it.