Retry sending emails a few times before giving up (#575)
Emails are tried once the normal way, and if they go through, there's no need
to create a job for them. If they fail with any exception, a Quartz job is
created allowing them to be retried later. Default is to retry every 5 minutes
and retry 3 times before giving up, but those defaults are parameters in the
SendEmailWithRetriesAsync method so they're easy to change.