Project invitation emails also register new users (#567)
The "You've been invited to join this project" email will link to the
new user registration page, with the user's name and email address
already filled in (and the JWT in a hidden form field). The password
field will be left blank; once the user fills it in, the user will be
registered. If the user did not change the email address in the
registration form, he/she can skip verifying the email. But we also
allow changing the email address, in which case the user will have to go
through the verification process.