First attempt, sometimes uses wrong Mercurial
First attempt at running tests with either Mercurial 3 or 6. It's
apparently not sufficient to reference the correct version of Chorus; we
need to clean the Mercurial folder out of the correct directory, or else
the most-recently-downloaded version gets used, due to the way Chorus
looks for the hg executable. Code for that will be in the next commit.