Fix updateLexEntryCount failing during SSR
The probable cause of the updateLexEntryCount failures is a race
condition between server-side rendering calling `fetch` and the
client-side rendered page calling the same `fetch` again, which
sometimes causes two serer processes to get into a race condition where
they check for the FlexProjectMetadata table row, don't find it, and
both try to create it with the same ProjectId value.

We catch the DbError on the server, and at the same time, we move the
updateLexEntryCount call onto the client (inside onMount) to avoid the
race condition ever happening in the first place.