Semi-working Login with Google button in UI
This implements a working "Login with Google" button if appropriate
GOOGLE_OAUTH_CLIENT_ID and GOOGLE_OAUTH_CLIENT_SECRET values have been
created in local.env. The login flow goes to the backend, then (at the
moment) simply returns the email address extracted from the JWT, so the
login button doesn't actually work yet. But it successfully executes the
oauth verification flow in ASP.NET Core, so the only thing left to do is
to extract the email address, check which user it belongs to, and log in
as that user.