Get all projects before updating any
If we try to use AsyncEnumerable here we get an error from Postgres:
NpgsqlOperationInProgressException. This is because the SELECT from the
Projects table is already running when we try to run UpdateLexEntryCount
(which does more DB operations on the same DbContext instance). So we'll
need to complete the SELECT before running the update.