Don't reuse HttpClient when updating all counts
The HgService class was designed to be instantiated once per request,
and so it ends up using the same HttpClient instance if it's called
multiple times in a loop. Since we're setting up for a temporary
migration and we're going to revert the code later, we might as well
create a new HttpClient instance each time here. This is wasteful of
socket connections, but it's temporary so we don't want to spend a lot
of time on a better solution.