Sign in with Google button component, first draft
This was copied directly from the HTML that Google first injects into
your page, meaningless CSS class names and all. I copied it exactly
as-is so that we could render their logo just as they want it rendered.
I'll tweak it to give the CSS classes more meaningful names, and to
remove any redundancies, but I'll leave it mostly unchanged apart from
renaming CSS classes.