Login with Google button (#597)
Instead of using the Google Sign In (GSI) Javascript library, we have 
the login process be handled by ASP.NET's OAuth packages. This allows us
to avoid pulling down 80+ kilobytes of Javascript that we have no
control over. We store the Google user ID (which we treat as an opaque
token) in the new GoogleId column of the Users table, and we look up a
GoogleId first before trying to match user accounts by email address.
This will allow us, in the future, to add UI for saying "Hey, I want to
connect my Google account to my LexBox account", or to add Facebook
and/or Microsoft logins with similar login flows.

---------

Co-authored-by: Kevin Hahn <hahn.kev@gmail.com>