Fix updateLexEntryCount failing during SSR (#596)
The cause of the updateLexEntryCount failures is when multiple calls to 
`fetch` for the same project would follow close on each others' heels.
This could sometimes cause two server processes to get into a race
condition where they check for the FlexProjectMetadata table row, don't
find it, and both try to create it with the same ProjectId value.

We catch the DbError on the server, and at the same time, we get rid of 
the derived lexEntryCount store on the client (which had been triggering
twice when the project page is loaded, thereby firing off two fetches
within milliseconds of each other) to avoid the race condition ever
happening in the first place. Instead, we simply ask for the
lexEntryCount correctly in the GQL request, and display a button on the
project page to refresh the count if it turns out to be out-of-date.

We also display a (very ugly) temporary button, marked as DevContent so 
only devs will see it, on the admin dashboard to update "all" (really
limit 100) lex entry counts for FLEx projects that don't have a valid
count yet.