Frontend can now inform backend of password strength
This may not be the best way to do it, though, as the backend would need
to trust the request coming from the frontend, and that could be
manipulated: someone with the password 1234 could craft a login request
with PasswordStrength set to 4 and it would be trusted. So we'll have to
scrap this approach and instead have the backend perform its own
independent calculation of password strength.