TEMP: Proof-of-concept for streaming hg commands
With this code, I can stream the command-runner.sh command output
through ASP.NET Core. It's only a proof-of-concept at the moment, and
there are some issues — like how streaming through `curl` at the command
line is doing the right thing, but streaming via an HTTP fetch still
seems to be buffered... *unless* I add a WriteAsJsonAsync() call that
dirties up the output. Still more to be solved...