Set Content-Type header so browser will stream
Without a Content-type, Firefox was waiting to see the entire content
before returning the results of the `.fetch()` call, in case... I don't
know in case of what. But if it knows ahead of time that the content
type is "text/plain; charset=utf-8" then it will stream the output to
.fetch one line at a time.

With this, I have everything I need to fix the problem.