Restore previous signature for Login method
We changed the method signature of LexAuthService.Login in order to pass
the User entity back, because we wanted to use it in calling
UserService.UpdatePasswordStrength. But now that UpdatePasswordStrength
just takes a GUID, we don't need the User entity from Login, so let's
put it back the way it was.