Frontend passes password strength in login request
While this could in theory be abused (someone could craft a JSON request
with password 1234 and claim it has password strength 4), in practice
it's not likely to matter, because we only set the password strength if
the password is valid, so you need to know someone's password to set the
password strength on their account. So we'll implement it as a way to
learn the password strength of people who don't use Send/Receive very
often (such as many admins, the ones we care most about).