Resetting password resets password srrength column
We might also choose to have the reset-password flow calculate password
strength and submit it; the considerations are the same as our login
flow, but it's slightly more API surface area where we'd be trusting
input from the frontend that we can't check. So for now I've decided to
just calculate password strength at login time, and not at password
reset time.