Have hg command runner stream its output (#616)
* Have hg command runner stream its output

If we delay writing anything to stdout, CloudFlare will time out the
connection after 100 seconds. But if we start writing to stdout and keep
writing to it, then CloudFlare will see an ongoing active connection and
won't time it out.

* Set Content-Type header so browser will stream

Without a Content-type, Firefox was waiting to see the entire content
before returning the results of the `.fetch()` call, in case... I don't
know in case of what. But if it knows ahead of time that the content
type is "text/plain; charset=utf-8" then it will stream the output to
.fetch one line at a time.

* Add 5-second timeout to hg verify

This ensures that hg verify won't lock the repo while it runs a
"checking files" step that we're never going to need (any repo
corruption we encounter is pretty much always due to an interrupted
transaction, which hg verify finds out about straight away).