Mark Email property on users as nullable
Now that we allow some users to not have emails, we should ensure our
Email property on User and LexAuthUser is marked nullable, so that the
C# compliler can help catch places where that matters.

We don't need to change the unique constraint on the DB column since
nulls aren't considered when dealing with unique constraints.