Remove CSS that does nothing
The badges in the project page are laid out using flexbox, not grid, so
the grid-column style does nothing. The `place-self-end` CSS class also
does nothing since it's not what it sounds like; it doesn't put the
badge at the end of the flexbox, it simply sets the align and justify
CSS props to "end" meaning the badge will be in the lower right-hand
corner of the box it's assigned by flexbox. (And since the badge uses
width 100% and height 100%, that makes no visual difference).