Improve text of bulk-add modal
We use the term "logins" consistently elsewhere instead of "usernames",
so let's make sure to mention that term here as well. But we also keep
the term "usernames" as that's more standard and easily-understandable
than "logins", at least to the experienced comptuer users who are the
only ones who will see this text.