UX improvement: final newline allowed in usernames list
Before this change, typing a final newline in the usernames list was
counting as having an empty username and this was being disallowed by
the validation code. (This would have also forbidden any blank lines in
the list). Now any empty strings in the usernames list are simply
removed before passing to the validation code, making for better UX.