SendVerifyAddressEmail throws if email is null
The only way we can get to this point with a null email address is if a
user has somehow managed to click on the "Send Verification Email"
button without having an email address. It's possible that we might make
mistakes in our UI in the future that could lead to this, so let's
handle this as an error rather than silenty doing nothing.