Add explicit foreign key on CreatedBy column
EF Core was already picking up the foreign key, but it's useful to be
able to specify it explicitly because we can specify the foreign key
deletion behavior as "RESTRICT", i.e. the database won't allow deleting
an admin user account until its CreatedBy users are reassigned to
another admin, or themselves deleted.