Simplify repo-moving script
By carefully adjusting the glob used, we can move only repos with names
that start with alphanumerics and that are two or more characters long.
This neatly separates repos from single-letter folders and also ensures
we won't move the `_____deleted_____` folder, or any other folder that
starts with an underscore for that matter.