Allow frontend to send strength score to backend
Three places in the frontend where password strength score will be
calculated and sent to the backend:

- When user registers
- When user resets a forgotten password
- When admin resets a user's password

This is in addition to when users log in with a username and password,
which already causes the strength score to be calculated (by the
backend). That will give us the password strength scores for most users
who registered before this feature was added. (Except for those who
don't log in for many months, in which case we'll probabaly end up
deactivating those inactive accounts).