Don't invite bulk-added user accounts
Instead, they are created with an email address and no username. If the
admin listed a display name with the email address then it is used,
otherwise the display name is set to be the same as the email address
(and the user will be able to change that later).