Fix S/R tests with Mercurial 6 (#716)
* Fix S/R tests with Mercurial 6

Latest Chorus version has some bugfixes to the Linux Mercurial package.

* Allow longer for S/R-after-reset test

hg push is timing out on develop server so it's hard to tell what's
going on as the "Thread abort unsupported on this platform" exception
swallows the original one. Temporarily triple the hg push timeout so we
have a chance of seeing what's going on.

* Always install python3-venv on self-hosted runner

The "Publish unit test results" step needs it. It's not strictly
necessary if we're running tests against Mercurial 3 since the Python 2
envrionment that setup-python installs does have virtualenv in it, but
it's simpler to just install it all the time.