Measure password strength when users log in or register (#614)
Password fields now show a password-strength meter providing instant
visual feedback on whether a user's password is good.

* Frontend can now inform backend of password strength

Three places in the frontend where password strength score will be
calculated and sent to the backend:

- When user registers
- When user resets a forgotten password
- When admin resets a user's password

This will give us the password strength scores for most users
who registered before this feature was added. (Except for those who
don't log in for many months, in which case we'll probably end up
deactivating those inactive accounts).

* Add NuGet package for password strength estimation

* Update password strength on backend if possible

When we receive a cleartext password, and we know that the password is
correct (the hash matches what we have), we can update the user's
password strength.

* Frontend passes password strength in login request

While this could in theory be abused (someone could craft a JSON request
with password 1234 and claim it has password strength 4), in practice
it's not likely to matter, because we only set the password strength if
the password is valid, so you need to know someone's password to set the
password strength on their account. So we'll implement it as a way to
learn the password strength of people who don't use Send/Receive very
often (such as many admins, the ones we care most about).

* Bulk-added users have password score of 0

Bulk-added users have a shared password, so no matter how complex it is,
their password score will always be set to 0, which will force a
password change in the future if they want to become a project manager.

* Simplify password strength calc in frontend login

We now calculate the password strength in the login helper function,
immediately before calling the backend API, rather than requiring the
Svelte component's score to be passed through to the helper function.

* Add a margin to password strength bar

* Fix progress bar colors not showing up

If you build the CSS name dynamically, like `progress-${color}`, it
doesn't get included in the final stylesheet because the build mechanism
doesn't recognize that it's used and strips it out. So we need to return
the entire string from passwordStrengthColor.

* Fix lint errors


---------

Co-authored-by: Tim Haasdyk <tim_haasdyk@sil.org>