Add proof-of-concept tool to make .fwdata files
Tool will be used to create .fwdata files on demand, allowing us to
implement several different features we might want on the backend.

Currently this is a separate binary so as not to pull in a bunch of
dependencies into the LexBoxApi project. But we could fold it into
LexBoxApi if we want to.