Make user typeahead safer for Playwright tests
The user typeahead had two debounces, one in the input and one in the
typeahead store. The debounce in the input was causing Playwright tests
to fail because they were filling in the input and expecting the form to
be valid immediately, but it wasn't valid until the debounce triggered.
The debounce on the input now serves no purpose; removing it.