Fix 404 during project reset test (#740)
* Remove race-condition window in ResetProject

* Remove race-condition window in FinishReset

* Faster SendReceiveAfterProjectReset test

Now that we've removed the race condition in ResetProject, we no longer
need to wait for the hgweb refresh interval in the middle of the
SendReceiveAfterProjectReset tests. (We still do need to wait for it at
the start of the test, after project creation).

* Delete invalid reset-project zip file contents

If the zip file uploaded during the reset project step had no .hg
directory anywhere inside it, we want to delete it entirely before
presenting the error message to the user. Otherwise we might leave quite
large amounts of useless data lying around in the repos volume.

* Add project reset 404 integration test

* Proper temp folder, restore metadata after reset

* Fix primary-key error on FlexProjectMetadata

* Address review comments

* simplify moving folders around for reset, introduce helper method to get a temp repo path

* Use ShouldAllBe for better error messages

* Keep test repo zip in Git instead of downloading it

This saves seeral seconds of test setup time.

* Move TusDotNetClient reference to correct location

This allows the project to build when MercurialVersion is set to 6.

* write tests for the integration fixture

* Make user typeahead safer for Playwright tests

The user typeahead had two debounces, one in the input and one in the
typeahead store. The debounce in the input was causing Playwright tests
to fail because they were filling in the input and expecting the form to
be valid immediately, but it wasn't valid until the debounce triggered.
The debounce on the input now serves no purpose; removing it.

* Make email-handling more robust in Playwright tests

Many times the Playwright code gets to the email page before the email
has actually shown up. So we need to make it wait for up to 10 seconds,
refreshing the email list periodically (because mailDev, at least,
doesn't auto-refresh: you have to click the refresh button).

* Just use template repo from orig location, no copy

* Nicer way to set TemplateRepoZip

* Simply use relative path for TemplateRepoZip

* Check email subjects and bump playwright timeout

* Fix IntegrationFixtureTests

---------

Co-authored-by: Tim Haasdyk <tim_haasdyk@sil.org>
Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>