Try to fix reset-project Playwright tests
We believe the failures in the Playwright tests are due to NFS caching
the directory entry, and thus not noticing that the directory has new
contents, resulting in Mercurial thinking the directory is still empty
for some time after the zip file upload has completed. It is possible to
force NFS to refresh its cache of a directory by doing opendir() and
closedir() on that directory. I don't know if C#'s Directory.GetFiles
method, which uses the openat() system call rather than opendir() and
closedir(), will have the same effect — but it's worth a try.