Fail reset-project test fast if NFS error persists
The reset-project E2E test has a built-in retry to try to work around
the 404 errors we'd been getting from hgweb. If those are cauesd by NFS
directory caching, then we should stop using a retry so that we get
those 404 errors fast — if we're going to get them, that is.