Create new invalidatedircache command in runner
This command will do nothing except run an `ls` on the requested project
code. The intent is to run this command during testing after a project
reset, in order to immediately invalidate the NFS dir cache on hgweb and
make the new state of the project available. It will likely not be needed
in production, because the only time the NFS cache has been relevant in
production (or staging) is when lexentrycount is run after a project
reset, and the `ls` is already baked into the lexentrycount command. But
in testing, it will sometimes be necessary to reset the NFS directory
cache in order to make the result of a project reset available right
away, which will allow the tests to stop being flaky.