Suppress expected error message when deleting repo
When deleting a repo, we call invalidatedircache afterwards to ensure
that the NFS cache picks up on the repo being gone. This means that `ls`
will not find the directory, and `cd` will fail too. So we silence
errors from `ls`, and exit before running `cd` if we're calling
invalidatedircache, because there's nothing left to do at that point.