Remove table view from MembersList
MembersList is going to move back to being a project-members component,
while the org page will use UsersTable to present the list of org
members. First step is removing the "Table view" toggle and the use of
UsersTable from the MembersList component, then we'll change its types
to be project-only.