Detect situation where empty repo zip uploaded
If FinishReset gets an upload of an *empty* Mercurial repo, then
WaitForRepoEmptyState could wait forever. To catch that situation 99% of
the time, we can simply check for the existence of `00changelog.i` in
the Mercurial repo store: that file is guaranteed to exist if the repo
has commits, so its absence means an empty repo.

To ensure belt-and-suspenders, we still need to implement a timeout in
the WaitForRepoEmptyState method, but this is a good start.