Slightly safer InvalidateDirCache
We now choose a random filename and check its existence before creating
it. In theory there's still a chance that another process could create
the file before we get a chance to, but in practice that's never going
to happen.

We also create a directory rather than a file because on the very, very
slim chance that another process creates a file between when we check
its existence and when we create it, the Directory.Create call will
fail. So it's minutely safer.