Ensure WaitForRepoEmptyState doesn't throw on timeout
I forgot that .NET code will throw OperationCanceledException if passed
a token that gets canceled. We don't want WaitForRepoEmptyState to throw
if its timeout is reached, we want it to return as if it was successful.
So we need to catch and suppress that exception.

Also passed the token into Task.Delay so that if the timeout is reached
during the delay, the delay will cancel immediately instead of running
to completion.