Try to fix reset-project Playwright tests (#789)
We believe the failures in the Playwright tests are due to NFS caching
the directory entry, and thus not noticing that the directory has new
contents, resulting in Mercurial thinking the directory is still empty
for some time after the zip file upload has completed. It is possible to
force NFS to refresh its cache of a directory by doing opendir() and
closedir() on that directory. We will add an `invalidatedircache`
command to our hgweb command runner which will run `ls` (which, among
other system calls, calls opendir() and closedir()), and call that
command after the lexbox pod changes the repo.

The other piece of the puzzle appears to be caching on the side of the
LexBox API pod's NFS client. According to the `nfs(5)` manpage, NFS
writes are cached until one of the following events occurs:

* Memory pressure forces reclamation of system memory resources.
* An application flushes file data explicitly with sync, msync, or fsync.
* An application closes a file with close.
* The file is locked/unlocked via fcntl.

So doing an `opendir/closedir` on the hgweb pod is not enough; we also
need to force the lexbox pod's NFS client to send its data to the NFS
server. I've implemented that by doing the equivalent of running
`mkdir random-name; rmdir random-name` inside the Mercurial repo after
the LexBox API pod touches it (e.g. a project reset or FinishReset
operation), followed by a call to `ls` on the hgweb side.