Improve display of usernames if missing
This is a change that was made to the admin dashboard on the develop
branch. Since the admin dashboard's user table was moved to the
UserTable component, we should replicate this change. (And remove it
from the admin dashboard, whre it is now a leftover fragment that no
longer compiles).