Overwrite local.env in `task setup` if it exists
Instead of using an external script to find a line in local.env, we
simply overwrite it if it already exists when running `task setup`. This
does mean that any modifications you've made to local.env will need to
be redone, but it allows `task setup` to be run twice if necessary.