Download large test repo once, then store in image (#850)
* Download large test repo once, then store in image

This allows us to skip the lengthy step of downloading elawa.zip from
Google Drive every time `skaffold delete` gets run, or the hg-repos
volume gets deleted.

* No longer unzip sena-3.zip into dev's Git worktree

There's no reason to unzip the sena-3.zip (or elawa.zip) files into the
dev's Git worktree any more. We had been doing this to populate the
repos volume when using Docker Compose, but now that we've switched
to Skaffold, this step is useless, so we'll remove it.