Change _deleteOrgUser to take user ID
Saves a lookup step and is more precise (no risk of deleting the wrong
user if we somehow end up with duplicate usernames or emails in the
database). There are no circumstances where we would need to delete a
user and *not* have that user's ID, so this is safer.