Remove TODO comments from DeleteOrg mutation
The ON DELETE CASCADE will take care of the Members table since it is
indeed a many-to-many table. And now that I've started implementing orgs
owning projects in the other branch, I know that the project ownership
will also be in a many-to-many table, so there's no need to do anything
about it here since ON DELETE CASCADE will also remove project ownership
(but not the projects themselves). Which means the transaction can also
go away, and we're left with very simple code.