Remove double-confirm from org deletion
It's unnecessary friction when testing org creation & deletion, and we
have a confirm dialog that requires you to type something so there's
actual thought required, which is enough of a barrier to prevent
accidental clicking having massive effects. Plus, we'll eventually be
moving to a soft-delete for orgs just like we have for projects, which
will make this unnecessary as soft-deletes are undoable.