Add OrgId index to migration
This was done by editing the migration by hand, which is only a good
idea if it has not yet been applied. But since running `skaffold delete`
will delete the DB and let it be re-created, this is not a major problem
since this PR is still in progress and has not yet been merged.