Org page (#800)
Org page is presented as tabs. One tab has a list of projects (net yet
implemented), another tab shows members (shown in table form similar to
admin dashboard), the third tab shows Settings, and the fourth tab is a
History tab that currently just shows the created and last-modified
dates. (In the future this will show "(username) was added to (org) on
(date)" and so on.)

For details on the commits in this PR, see the PR page.

---------

Co-authored-by: Tim Haasdyk <tim_haasdyk@sil.org>