Proof-of-concept for bulk-add users on admin dash
This isn't great UI so we'll almost certainly do it differently, but
it's one possible way this could be done. The feature is accessed from
the Settings tab on the org page; when you click on it, you're taken to
the admin dashboard where the bulk-add button is now visible.